validator for itemRemoved added

+++ Controllers/OrderController.js
@@ -4,6 +4,7 @@ const OrderService = require('../Services/OrderService')
 const JSONResponseService = require('../Services/JSONResponseService')
 
 const orderValidate = require('../Validators/newOrderValidator.json')
+const idRemoveValidator = require('../Validators/removeOrderItem.json')
 const Ajv = require('ajv')
 const addFormats = require('ajv-formats')
 
@@ -18,7 +19,6 @@ let createNewOrder = (req, res) => {
             postcode: req.body.postcode,
             email: req.body.email,
         }
-
         const newOrderValidate = ajv.compile(orderValidate)
         const valid = newOrderValidate(order)
         if (valid) {
@@ -48,16 +48,27 @@ let removeOrderItem = (req, res) => {
             orderId: ObjectId(req.body.orderId),
             menuItemId: req.body.menuItemId,
         }
-            const removeOrder = await OrderService.removeOrderItem(db, item)
-            if(removeOrder.modifiedCount === 1) {
-                let response  = JSONResponseService.generateSuccessResponse()
-                response.message = "All dishes of this quantity successfully deleted from order"
+        const idValidator = ajv.compile(idRemoveValidator)
+        const valid = idValidator(item)
+        if (valid) {
+            try {
+                const removedOrder = await OrderService.removeOrderItem(db, item)
+                if (removedOrder.modifiedCount === 1) {
+                    let response  = JSONResponseService.generateSuccessResponse()
+                    response.message = "Item removed"
+                    response.data = removedOrder.ops
+                    return res.json(response)
+                }
+            } catch (e) {
+                let response = JSONResponseService.generateFailureResponse()
+                response.message = "Database request failed"
                 return res.json(response)
             }
-            let response = JSONResponseService.generateFailureResponse()
-            response.message = "Item was not deleted from order"
-            return res.json(response)
-        } )
+        }
+        let response = JSONResponseService.generateFailureResponse()
+        response.message = "Validator failed"
+        return res.json(response)
+    })
     }
